refactor(#45): ProfileData data class 패키지 이동
